requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.6.1
dnspython==2.4.2
python-dotenv==1.0.0
pyjwt==2.8.0
gunicorn==21.2.0